moves test to interactions_test
